upvote og downvote

+++ Model/Bruger.cs
@@ -3,14 +3,14 @@
     public class Bruger
     {
         //Konstrukter
-        public Bruger (long brugerID, string brugernavn)
+        public Bruger (int brugerID, string brugernavn)
         {
             this.BrugerID = brugerID;
             this.BrugerNavn = brugernavn;
         }
 
         //Felter
-        public long BrugerID { get; set; }
+        public int BrugerID { get; set; }
         public string BrugerNavn{ get; set; }
     
        //Tom konstrukter
