+++ Model/Kommentar.cs
@@ -3,7 +3,7 @@
     public class Kommentar
     {
         //Konstrukterer
-        public Kommentar(long kommentarID, string tekst, Bruger bruger, long upvotes, long downvotes, DateTime dato = new DateTime())
+        public Kommentar(int kommentarID, string tekst, Bruger bruger, int upvotes, int downvotes, DateTime dato = new DateTime())
         {
             this.KommentarID = kommentarID;
             this.Tekst = tekst;
@@ -12,7 +12,7 @@
             this.DownVotes = downvotes;
             this.Dato = dato;
         }
-        public Kommentar(long kommentarID , string tekst, Bruger bruger, long upvotes, long downvotes)
+        public Kommentar(int kommentarID , string tekst, Bruger bruger, int upvotes, int downvotes)
         {
             this.KommentarID = kommentarID;
             this.Tekst = tekst;
@@ -22,11 +22,11 @@
         }
 
         //Felter
-        public long KommentarID { get; set; }   
+        public int KommentarID { get; set; }   
         public string Tekst { get; set; }
         public Bruger Bruger { get; set; }
-        public long UpVotes { get; set; }
-        public long DownVotes { get; set; }
+        public int UpVotes { get; set; }
+        public int DownVotes { get; set; }
         public DateTime Dato { get; set; } = DateTime.Now;
